Break out functions from content_based.py
Move the functions to their own library, and replace `vect_mult` with
`np.dot` which performs the same function.
